gluster-blockd: treat delete success if block doesn't exist
problem:
say because of some reasons create failed to config on node, so it will
log CONFIGFAIL in the metafile. At the time of delete since config
failed it is obvious that block will not exist. since delete cannot find
it now it fails to delete hence returning -1.

solution:
treat delete success if there is no block with given name on that node.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>