build: pass "--with systemd" to rpmbuild if needed
In case ./configure detects that the environment is using systemd, the
"make install" step will copy the gluster-block.service file to the
right location. However, the RPMs do not include this file by default,
so adding a "--with systemd" to rpmbuild/mock is needed (until there is
some form of auto-detection in the .spec).

Change-Id: Id64fa664f5b2a3d3c4675b2ad7a54d415a54281a
Signed-off-by: Niels de Vos <ndevos@redhat.com>