Prevent crashes when errMsg is not set
When reply->out is NULL, gluster-blockd is crashing in serializing the response.
Unfortunately the function where we can handle this error is auto-generated, so
I added catch-all errors with a generic-message if errMsg is not set.

Also improved errMsg in somecases where it is not setting by collecting the
error messages from all the config-nodes.

Change-Id: I706d8af9e8e6140d21c0d1268ae57dd5f364aa54
Signed-off-by: Pranith Kumar K <pkarampu@redhat.com>