Only when RPC is sent move status to CONFIGFAIL
Problem:
When the config node is given as an address that has no route,
connection to the node fails with EHOSTUNREACH(No route to Host).
Since the code doesn't treat it as special errno, the config
ends up updating its status as CONFIGFAIL and as part of cleanup
it attempts to delete it, but fails again because the host is
not reachable. So the block file lingers on in that state forever.

Fix:
Irrespective of errno, unless we don't send an rpc we know that
we don't need to move the status to CONFIGFAIL, so get that info
and act accordingly irrespective of the errno.

Change-Id: Ie3fb9413f089a4875b5612c60ed3d65bc5b5bba1
Signed-off-by: Pranith Kumar K <pkarampu@redhat.com>