build: do not require 'git' to find the version
While looking into packaging gluster-block for Fedora, I noticed that
'git' is used to determine the version. The downloads from GitHub are
not archives that have been created with 'make dist', and hence
./autogen.sh is still needed to be run.

./autogen.sh generates all the needed bits, including the ./configure
script that tries to figure out the version. If ./configure runs in a
non-git directory, no version can be found.

In order to solve this, I'm adding a VERSION file in the root of the
project. If this file exists, it is used to fetch the version, instead
of trying to run git. The basic usage for RPM packaging in distributions
can then be:

  echo %{version} > VERSION
  ./autogen.sh
  %configure

Note that the VERSION file is not needed for building from a git
repository. However, this file can get packaged in the 'make dist'
tarball so that building from the tarball does not require git.

Change-Id: Ied378c7071ee4a108a1e946dccbb7f223b7aeb9f
Updates: #25
Signed-off-by: Niels de Vos <ndevos@redhat.com>