tcmu: backstore attribute set cmd_time_out=0
currently, runner is not supporting the non-zero cmd timeouts (cmd_time_out !=0)

The tcmu cmd timer just fails the kernel command after cmd_time_out seconds,
leaving the runner command running. When it is zero, it means do not run the
timer, so both the kernel and runner wait for the command to complete.

This patch sets the backstore attribute cmd_time_out=0

Change-Id: Ib7a80c82d0aafe345ff4e79ce88dc92848306f4e
Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>