delete/rollback: check for matching gbid before deleting a block
Currently we are just checking for blockname in targetcli list output, if it
exist we will delete the backend.

We do not check if block belongs to a given volume (say vol1), because of which
a rollback on a failure to create block with same name, on same node but on a
different volume (say vol2) deletes block backend belonging to vol1.

Solution:
only if blockname + gbid matches continue delete.

Change-Id: I396c228dad3cf4f51cc6676a266b837bdf6040c7
Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>