saveconfig: only alter given block configuration
We use saveconfig in all the remote calls i.e. create/delete/modify/replace.

Currently, if we try to operate on a given block, the consequent saveconfig
will result in rewriting the whole node configuration again as per the current
state of targets in the machine, this could result in override of currently
existing but inactive blocks configuration.

This patch support saving configuration per block. i.e. only the configuration
of the block that is requested is altered, rest all will be untouched.

Thus, this patch can avoid overriding of the currently unloaded configurations.

Fixes: #70
Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>