rpc: glfs-operations: update api call for ftruncate
The updated gfapi for glusterfs fop ftruncate has pre/post
attributes in the function signature. But this api call
inside gluster-block is not update, so update the
api call for ftruncate to resolve build errors.

Make sure 'config.h' is added in build properly

Notice that, this is required for building gluster-block
with glusterfs 'master' branch. As glusterfs glfs-api signature
and the development header path are being modified in master,
there may be more things in plate before next version release,
and hence keeping the flag as 'NOT_NEW' for current stack.

Fixes: #75
Signed-off-by: Bhumika Goyal <bgoyal@redhat.com>
Signed-off-by: Amar Tumballi <amarts@redhat.com>