gluster-block: load-balance support
Problem:
-------
If we have say 4 blocks with HA 3, there is possibility that for all the
4 devices when logged at the initiator, active IO paths connect to only one
node out of 3, this is because we do not decide the priority of the target
portals for a device on the target side and set priority to constant
(all paths prio 1) at the initiator side.

So there is no guarantee of load balance among 3 nodes.

Solution:
-------
By setting one portal to active optimised (AO) alua access state and other to
active non-optimised (ANO), the priority of portal which is assigned to AO is
set to 50 and for ANO it is 10. So portals assigned to AO will be considered
to be active paths.

We maintain the data about the priority paths in each glusterfs volume
as xattrs on '/prio.info' file

While we create a block, counter for respective prio portal is incremented

Ex:
$ for i in {1..4}; do \
  gluster-block create sample/block${i} ha 3 \
   192.168.124.149,192.168.124.162,192.168.124.184 1GiB; \
done

$ getfattr -m. -d -e text /mnt/block-meta/prio.info
[...]
user.block.192.168.124.149="2\0..."
user.block.192.168.124.162="1\0..."
user.block.192.168.124.184="1\0..."

While we delete a block, counter for respective prio portal is decremented

$ gluster-block delete sample/block4

$ getfattr -m. -d -e text /mnt/block-meta/prio.info
[...]
user.block.192.168.124.149="1\0..."
user.block.192.168.124.162="1\0..."
user.block.192.168.124.184="1\0..."

Fixes: #82
Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>