gluster-block: fill errMsg and delete metafile if EntryCreate fail
Reason for crash:
When a call to glusterBlockCreateEntry() fails, we goto exist label, and
eventually call blockCreateCliFormatResponse() with arg errMsg containing NULL.
Now inside blockCreateCliFormatResponse() if errMsg is NULL, we might treat as
success and continue, which causes dereferencing savereply (NULL pointer).

Fix:
Make sure errMsg is not NULL in case of failures.

Tested-by: Xiubo Li <xiubli@redhat.com>
Reviewed-by: Xiubo Li <xiubli@redhat.com>
Reviewed-by: Pranith Kumar K <pkarampu@redhat.com>
Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>