Script to wait for bricks to come online
Problem:
When a node is rebooted, by the time tcmu-runner starts there is no
guarantee that the bricks of the volume on which the block-volumes
are created will be online. This leads to a case where doing open
on the block files leads to open failing with ENOTCONN leading
to containers with this as pvc are in failed state.

Fix:
A script is added which will be added as a unit file dependency for
gluster-blockd which makes sure that the bricks are online for the
block-hosting-volume preventing this situation.

Reviewed-by: ndevos@redhat.com
Reviewed-by: pkalever@redhat.com
Tested-by: pkalever@redhat.com