capabilities: read capabilities is one-time job
I do not see any value in reading the capabilities every time a command
kicks in, in fact this is a major bug.

Say we are running a daemon of version X, in case if the gluster-block package
is upgraded to version Y, then X capabilities file will be overridden by
Y version capability file, even before restarting the daemon the gluster-blockd
process of version X can return wrong capabilities of version Y.

Also, we are adding delay in the gluster-blockd overall response by
unnecessarily reading the capabilities on every command from the file.

Tested-by: Xiubo Li <xiubli@redhat.com>
Reviewed-by: Xiubo Li <xiubli@redhat.com>
Reviewed-by: Pranith Kumar K <pkarampu@redhat.com>
Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>
Signed-off-by: Xiubo Li <xiubli@redhat.com>