wait-for-bricks: Use ps instead of pidof, improve performance
1) 'pidof' command may not exist in all setups, so changing that
to use 'ps' command which is available more often. If 'ps' command is
not found in the setup it prints that the command is not available
in the setup in the error log.

2) For finding if 10 block volumes are up and running it was taking 9 seconds
when all the bricks are up. I changed the script to reduce the number of
gluster commands that are executed to get this time to 1 seconds.

Tested-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>
Reviewed-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>
Reviewed-by: Niels Devos <ndevos@redhat.com>
Signed-off-by: Pranith Kumar K <pkarampu@redhat.com>