[WIP] correct some rpc socket setup bits
Creating the socket is not needed for svcunix_create(), it does all the
work for us (and prevents "address already in use" errors).

For svctcp_create(), the socket is expected on port 24010 and hence
needs to be setup completely with bind() and listen() before it can be
used. Registering the RPC-program should make it possible to use a
dynamically assigned port, but the gluster-block CLI expects it on the
static port.

This also fixes a warning when building the RPM from 'make dist' tarball
as the 22 June 2017 was a Thursday, nt Tuesday.

Change-Id: If3a6b7527399dd0a5a16f4273efdd607617289de