extras: add a script for upgrade time activities
This script is a placeholder to run any upgrade time activities.

For now we generate the gluster block target configuration as part of this script.
This will only continue to run till its first successful exit, when it
successfully finish the first run post upgrade, it will maintain/touch a status
file under /var/lib/gluster-block, if the status file exist then the upgrade
activity script is skipped.

This script reuse some logic from wait-for-bricks.sh, so want to extend the
credits to Pranith on that note.

Reviewed-by: Niels Devos <ndevos@redhat.com>
Reviewed-by: Amar Tumballi <amarts@redhat.com>
Reviewed-by: Raghavendra Talur <rtalur@redhat.com>
Reviewed-by: Xiubo Li <xiubli@redhat.com>
Tested-by: Xiubo Li <xiubli@redhat.com>
Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>