arpg: provide an option to set gluster hostname and port
The goal is to keep the changes to minimal from how gluster-block
project behaves today. Only if the hostname/IP is passed to g-b
daemon (which would happen in special cases, where GBD is in
separate pod than that of glusterfs pod), the behavior of fetching
volfile from given volume comes to picture.

Fixes: #136
Signed-off-by: Amar Tumballi <amarts@redhat.com>