coverity: fix pointless_string_compare issues (#171)
LOG() macro was designed to compare for specific strings, and based
on the result, log to different log files. This was done using
'strcmp()', but was using static strings. It made sense to compare
the static string directly as they resolve to same pointer while
compiling.

With this fix, was able to resolve 58 coverity issues in one shot.

Signed-off-by: Amar Tumballi <amarts@redhat.com>
Reviewed-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>
Reviewed-by: Xiubo Li <xiubli@redhat.com>