inotify: watch the directory not just file
Editors (such as vim, nano ..) follow different approaches to save conf file.
The two commonly followed techniques are to overwrite the existing file, or to
write to a new file (.swp, .tmp ..) and move it to actual file name later. In
the later case, the inotify fails, because the file it's been intended to watch
no longer exists, as the new file is a different file with just a same name.

To handle both the file save approaches mentioned above, it is better we watch
the directory and filter for MODIFY events.

Signed-off-by: Xiubo Li <xiubli@redhat.com>