daemon: convert cli and remote threads to individual processes
Problem:
-------
When using TIRPC we have seen some contention with poll fd's, this was
because we are calling svc_run() in both the cli and remote threads.

Probably an issue in TIRPC as the current model worked well for us with
glibc sun-rpc implementation.

Solution:
--------
Convert cli and remote threads to individual processes

How to compile ?
----------------
On systems having glibc sun-rpc: (F27/Centos/RHEL)
$ ./autogen.sh && ./configure --enable-tirpc=no

On systems having tirpc: (F28 & above)
$ ./autogen.sh && ./configure

Thanks to Xiubo for debugging this issue along.

Closes: #57
Closes: #165
Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>
Reviewed-by: Xiubo Li <xiubli@redhat.com>
Reviewed-by: Amar Tumballi <amarts@redhat.com>