dyn-config: try to read it again in case the pagecache is not flushed out
Problem:
-------
We are hitting one case that when trying to change the config file and
close it, all the uncomment options will be fall back to the DEFAULT values.

Resolution:
----------
In-case if the editor (vim) follows write to a new file (.swp, .tmp ..)
and move it to actual file name later. There is a window, where we will
encounter one case that the file data is not flushed to the disk, so in
another process(here) when reading it will be empty.

Let just wait and try again.

Signed-off-by: Xiubo Li <xiubli@redhat.com>
Reviewed-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>