delete: use gbid instead of block_name to do the saveconfig check
This will be one problem if we create 2 block targets by using the
same NAME in 2 different volumes. The second time it will fail and
won't be loaded and no any info in the saveconfig.json, but it will
still could find it which is the StorageObject created by the first
creation, the the rollback will fail like:

IQN: -
PORTAL(S): -
ROLLBACK FAILED ON: 192.168.195.164
RESULT: FAIL

But the rollback is actually successful. After this fixing it will
be like:

failed to configure on 192.168.195.164 create StorageObject failed
RESULT:FAIL

Signed-off-by: Xiubo Li <xiubli@redhat.com>
Reviewed-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>